+++ tests/contrib/test_pydantic.py
@@ -14,6 +14,7 @@ from tests.testmodels import (
     Team,
     Tournament,
     User,
+    json_pydantic_default,
 )
 from tortoise.contrib import test
 from tortoise.contrib.pydantic import (
@@ -1185,6 +1186,7 @@ class TestPydantic(test.TestCase):
                 "data_null": None,
                 "data_default": {"a": 1},
                 "data_validate": None,
+                "data_pydantic": json_pydantic_default.model_dump(),
             },
         )
         ret1 = creator.model_validate(json_field_1_get).model_dump()
@@ -1196,6 +1198,7 @@ class TestPydantic(test.TestCase):
                 "data_null": None,
                 "data_default": {"a": 1},
                 "data_validate": None,
+                "data_pydantic": json_pydantic_default.model_dump(),
             },
         )