+++ CHANGELOG.rst
@@ -23,6 +23,7 @@ Added
 ^^^^^
 - Add POSIX Regex support for PostgreSQL and MySQL (#1714)
 - support app=None for tortoise.contrib.fastapi.RegisterTortoise (#1733)
+- Added ``table_name_generator`` Meta attribute to allow dynamic table name generation (#1770)
 
 
 0.21.6